crimson/net: move message read path from ProtocolV2 to Protocol
Also move socket shutdown ownership to Protocol at READY/open.
