+++ src/crimson/net/Protocol.h
@@ -47,7 +47,7 @@ class Protocol {
 
   virtual void notify_out() = 0;
 
-  virtual void notify_out_fault(std::exception_ptr) = 0;
+  virtual void notify_out_fault(const char *where, std::exception_ptr) = 0;
 
 // the write state-machine
  public:
@@ -107,16 +107,7 @@ class Protocol {
 
   void set_out_state(const out_state_t &new_state);
 
-  seastar::future<> wait_out_exit_dispatching() {
-    if (out_exit_dispatching) {
-      return out_exit_dispatching->get_shared_future();
-    }
-    return seastar::now();
-  }
-
-  void notify_keepalive_ack(utime_t keepalive_ack);
-
-  void notify_ack();
+  seastar::future<> wait_io_exit_dispatching();
 
   void requeue_out_sent_up_to(seq_num_t seq);
 
@@ -132,18 +123,8 @@ class Protocol {
     return is_out_queued() || !out_sent_msgs.empty();
   }
 
-  void ack_out_sent(seq_num_t seq);
-
-  void set_last_keepalive(clock_t::time_point when) {
-    last_keepalive = when;
-  }
-
   seq_num_t get_in_seq() const {
     return in_seq;
-  }
-
-  void set_in_seq(seq_num_t _in_seq) {
-    in_seq = _in_seq;
   }
 
   ChainedDispatchers& dispatchers;
@@ -170,6 +151,12 @@ class Protocol {
       bool require_ack);
 
   void notify_out_dispatch();
+
+  void ack_out_sent(seq_num_t seq);
+
+  seastar::future<> read_message(utime_t throttle_stamp, std::size_t msg_size);
+
+  void do_in_dispatch();
 
   crimson::common::Gated gate;
 
@@ -207,6 +194,8 @@ class Protocol {
   /*
    * in states for reading
    */
+
+  std::optional<seastar::shared_promise<>> in_exit_dispatching;
 
   /// the seq num of the last received message
   seq_num_t in_seq = 0;
